拆分请求和响应方法
